fix: return type in users api functions
