+++ src/api/friends/users.ts
@@ -1,9 +1,10 @@
 import axios from 'axios'
 import { errorHandler, requestWithRefreshToken } from '../api'
 import { getCookie } from '@src/utils/cookie'
+import { type IAddFriendResponse, type IAllUsersResponse } from '@src/types/api'
 
-export const getAllUsers = async () => {
-  const getAllUsersRequest = async () => {
+export const getAllUsers = async (): Promise<IAllUsersResponse> => {
+  const getAllUsersRequest = async (): Promise<IAllUsersResponse> => {
     const res = await axios
       .get('/users', { headers: { Authorization: `Bearer ${getCookie('accessToken')}` } })
       .catch(errorHandler)
@@ -14,8 +15,8 @@ export const getAllUsers = async () => {
   return await requestWithRefreshToken(getAllUsersRequest)
 }
 
-export const addFriend = async (id: number) => {
-  const addFriendRequest = async () => {
+export const addFriend = async (id: number): Promise<IAddFriendResponse> => {
+  const addFriendRequest = async (): Promise<IAddFriendResponse> => {
     const res = await axios
       .post(
         `/users/${id}`,
